add pagination in all request reception

+++ src/modules/controllers/reception.controllers.js
@@ -9,6 +9,7 @@ const tokenVerify = (token) => {
 
 module.exports.getAllReceptions = (req, res) => {
   const { headers, query } = req;
+
   if (query.hasOwnProperty('limit')
     && query.hasOwnProperty('offset')
     && headers.hasOwnProperty('authorization')
@@ -16,6 +17,7 @@ module.exports.getAllReceptions = (req, res) => {
     && query.offset
     && headers.authorization) {
     const tokenParse = tokenVerify(headers.authorization);
+
     const limit = +(query.limit);
     const offset = +(query.offset);
 
@@ -34,26 +36,40 @@ module.exports.getAllReceptions = (req, res) => {
 };
 
 module.exports.createNewReception = (req, res) => {
-  const { body, headers } = req;
-
-  const tokenParse = tokenVerify(headers.authorization);
+  const { body, headers, query } = req;
 
   const flag = body.hasOwnProperty('nameUser')
     && body.hasOwnProperty('nameDoctor')
     && body.hasOwnProperty('date')
     && body.hasOwnProperty('complaint')
-    && body.nameUser !== ''
-    && body.nameDoctor !== ''
-    && body.date !== ''
-    && body.complaint !== '';
+    && query.hasOwnProperty('limit')
+    && query.hasOwnProperty('offset')
+    && headers.hasOwnProperty('authorization')
+    && body.nameUser
+    && body.nameDoctor
+    && body.date
+    && body.complaint
+    && query.limit
+    && query.offset
+    && headers.authorization;
 
   if (flag) {
+    const tokenParse = tokenVerify(headers.authorization);
+
     body["idUser"] = tokenParse._id;
     const reception = new Receptions(body);
 
+    const limit = +(query.limit);
+    const offset = +(query.offset);
+
+    const startInd = offset * limit;
+
     reception.save(body).then(() => {
-      Receptions.find({ idUser: tokenParse._id }, ['nameUser', 'nameDoctor', 'date', 'complaint']).then(result => {
-        res.send({ data: result });
+      Receptions.find({ idUser: tokenParse._id }).then(resultLength => {
+        const lengthResult = resultLength.length;
+        Receptions.find({ idUser: tokenParse._id }, ['nameUser', 'nameDoctor', 'date', 'complaint']).skip(startInd).limit(limit).then(result => {
+          res.send({ data: result, length: lengthResult });
+        });
       });
     });
   } else {
@@ -62,17 +78,29 @@ module.exports.createNewReception = (req, res) => {
 };
 
 module.exports.editReception = (req, res) => {
-  const { body, headers } = req;
+  const { body, headers, query } = req;
 
   if (body.hasOwnProperty('_id')
+    && query.hasOwnProperty('limit')
+    && query.hasOwnProperty('offset')
     && headers.hasOwnProperty('authorization')
-    && body._id !== ''
-    && headers.authorization !== '') {
+    && body._id
+    && query.limit
+    && query.offset
+    && headers.authorization) {
     const tokenParse = tokenVerify(headers.authorization);
 
+    const limit = +(query.limit);
+    const offset = +(query.offset);
+
+    const startInd = offset * limit;
+
     Receptions.updateOne({ _id: body._id }, body).then(() => {
-      Receptions.find({ idUser: tokenParse._id }, ['nameUser', 'nameDoctor', 'date', 'complaint']).then(result => {
-        res.send({ data: result });
+      Receptions.find({ idUser: tokenParse._id }).then(resultLength => {
+        const lengthResult = resultLength.length;
+        Receptions.find({ idUser: tokenParse._id }, ['nameUser', 'nameDoctor', 'date', 'complaint']).skip(startInd).limit(limit).then(result => {
+          res.send({ data: result, length: lengthResult });
+        });
       });
     });
   } else {
@@ -84,14 +112,26 @@ module.exports.deleteReception = async (req, res) => {
   const { headers, query } = req;
 
   if (query.hasOwnProperty('_id')
+    && query.hasOwnProperty('limit')
+    && query.hasOwnProperty('offset')
     && headers.hasOwnProperty('authorization')
     && query._id
+    && query.limit
+    && query.offset
     && headers.authorization) {
     const tokenParse = tokenVerify(headers.authorization);
 
+    const limit = +(query.limit);
+    const offset = +(query.offset);
+
+    const startInd = offset * limit;
+
     Receptions.deleteOne({ _id: query._id }).then(() => {
-      Receptions.find({ idUser: tokenParse._id }, ['nameUser', 'nameDoctor', 'date', 'complaint']).then(result => {
-        res.send({ data: result });
+      Receptions.find({ idUser: tokenParse._id }).then(resultLength => {
+        const lengthResult = resultLength.length;
+        Receptions.find({ idUser: tokenParse._id }, ['nameUser', 'nameDoctor', 'date', 'complaint']).skip(startInd).limit(limit).then(result => {
+          res.send({ data: result, length: lengthResult });
+        });
       });
     });
   } else {
